Create SyncUpdateWorker, WorkerUtils and WorkManagerSynUpdateRepository
This commit introduces the `SyncUpdateWorker` class, which is responsible for synchronizing updates in the background. The `WorkerUtils` class is added to provide utility functions for worker tasks, specifically for creating update notifications. The `WorkManagerSynUpdateRepository` is implemented to schedule the `SyncUpdateWorker` to run periodically using WorkManager, ensuring updates are checked and applied regularly.
- Creates `SyncUpdateWorker` to handle background synchronization of updates.
- Adds `WorkerUtils` for worker-related utilities, including notification creation.
- Implements `WorkManagerSynUpdateRepository` to manage the scheduling of the `SyncUpdateWorker` using WorkManager.
- Sync new files and clear garbage data.
- Check for the app version and notify if update is needed.
- Sync new ads from remote.
- Schedule `SyncUpdateWorker` to run every 6 hours.
- Add notification channel for update notification.