Update download method to support sendId
The download method in `FileDownloaderRepository` now accepts an optional `sendId` parameter.
The `FileEntity` model has been updated to include a `sendid` property.
The `HomeViewModel` now passes the `sendid` when initiating a file download.
The file downloader now send `sendid` in the request header if provided and if the request is for telebirr url.