Group files by type and folder and show type as sticky header.
The `DownloadList` composable now groups `DownloadItemData` by their media type or file type and by folder.
It also shows the media or file type as a sticky header.
The file group by is changed from argument to group by in the `DownloadList` composable.
The sticky header is shown only if there is no filter selected.